refactor: Migrate analysis models to pydantic v2 (#1306)
