+++ BALSAMIC/constants/workflow_params.py
@@ -10,98 +10,98 @@ VARCALL_PARAMS = {
 VCF_DICT = {
     "tnscope_umi": {
         "mutation": "somatic",
-        "type": "SNV",
+        "mutation_type": "SNV",
         "analysis_type": ["single", "paired"],
         "sequencing_type": ["targeted"],
         "workflow_solution": ["Sentieon_umi"],
     },
     "tnscope": {
         "mutation": "somatic",
-        "type": "SNV",
+        "mutation_type": "SNV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["wgs"],
         "workflow_solution": ["Sentieon"],
     },
     "dnascope": {
         "mutation": "germline",
-        "type": "SNV",
+        "mutation_type": "SNV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["Sentieon"],
     },
     "manta": {
         "mutation": "somatic",
-        "type": "SV",
+        "mutation_type": "SV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "cnvkit": {
         "mutation": "somatic",
-        "type": "CNV",
+        "mutation_type": "CNV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted"],
         "workflow_solution": ["BALSAMIC"],
     },
     "vardict": {
         "mutation": "somatic",
-        "type": "SNV",
+        "mutation_type": "SNV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted"],
         "workflow_solution": ["BALSAMIC"],
     },
     "manta_germline": {
         "mutation": "germline",
-        "type": "SV",
+        "mutation_type": "SV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "haplotypecaller": {
         "mutation": "germline",
-        "type": "SNV",
+        "mutation_type": "SNV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted"],
         "workflow_solution": ["BALSAMIC"],
     },
     "dellysv": {
         "mutation": "somatic",
-        "type": "SV",
+        "mutation_type": "SV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "tiddit": {
         "mutation": "somatic",
-        "type": "SV",
+        "mutation_type": "SV",
         "analysis_type": ["single", "paired"],
         "sequencing_type": ["wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "dellycnv": {
         "mutation": "somatic",
-        "type": "CNV",
+        "mutation_type": "CNV",
         "analysis_type": ["single", "paired"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "ascat": {
         "mutation": "somatic",
-        "type": "CNV",
+        "mutation_type": "CNV",
         "analysis_type": ["paired"],
         "sequencing_type": ["wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "cnvpytor": {
         "mutation": "somatic",
-        "type": "CNV",
+        "mutation_type": "CNV",
         "analysis_type": ["single"],
         "sequencing_type": ["wgs"],
         "workflow_solution": ["BALSAMIC"],
     },
     "svdb": {
         "mutation": "somatic",
-        "type": "SV",
+        "mutation_type": "SV",
         "analysis_type": ["paired", "single"],
         "sequencing_type": ["targeted", "wgs"],
         "workflow_solution": ["BALSAMIC"],
