+++ BALSAMIC/utils/metrics.py
@@ -4,4 +4,4 @@ from BALSAMIC.models.metrics import MetricValidation
 
 def validate_qc_metrics(metrics: dict) -> dict:
     """Returns a set of validated QC metrics."""
-    return MetricValidation(metrics=metrics).dict()["metrics"]
+    return MetricValidation(metrics=metrics).model_dump()["metrics"]
